requirements.txt
streamlit
flask
opencv-python
easyocr
sympy
pillow
numpy
python-dotenv
google-generativeai
requests